Some str

+++ forecastModel.py
@@ -1,7 +1,8 @@
+from unittest import result
 import pandas as pd
 from fileWork import *
 from sklearn.model_selection import train_test_split
-from sklearn.linear_model import SGDClassifier, LogisticRegression
+from sklearn.linear_model import SGDClassifier, LogisticRegression, Lasso
 from sklearn.preprocessing import StandardScaler
 from sklearn.pipeline import make_pipeline
 import pickle
@@ -10,18 +11,45 @@ import pickle
 class Model:
 
     def __init__(self, df, mainLogFilename = "./results/log/main.log", 
-                    logFilename="./results/log/logit.log",
-                    logitModelFilename = "./results/models/logit.pickle"):
+                    logitLogFilename="./results/log/logit.log",
+                    logitModelFilename = "./results/models/logit.pickle",
+                    fwLassoModel = "./results/models/lasso.pickle",
+                    fwSGDModel = "./results/models/sgd.pickle",
+                    lassoLogFilename = "./results/log/lasso.log",
+                    sgdLogFilename = "./results/log/sgd.log",
+                    pathesDict=None):
         """
         Class for build model to forecast data and
         other helpfull statistic methods
+
+        Args:
+            mainLogFilename(string, default: "./results/log/main.log"):
+                path to main.log file
+            logFilename(string, default: "./results/log/logit.log"):
+                path to logit.log file
+            logFilename(string, default: "./results/log/logit.log"):
+                 Env(dict, default: None): most easy way to update pathes
+                for log file. Just build .env file (like in .env_example)
+                and put this env on dict with param name. 
         """
         self.df = df
         self.logitModelFilename = logitModelFilename
-        self.fwMain = FileWork(mainLogFilename)
-        self.fwLogit = FileWork(logFilename)
-        self.fwLogitModel = FileWork()
-
+        if pathesDict is None:
+            self.fwMain = FileWork(mainLogFilename)
+            self.fwLogit = FileWork(logitLogFilename)
+            self.fwLogitModel = FileWork(logitModelFilename)
+            self.fwLassoModel = FileWork(fwLassoModel)
+            self.fwSGDModel = FileWork(fwSGDModel)
+            self.fwlassoLogFilename = FileWork(lassoLogFilename)
+            self.fwsgdLogFilename = FileWork(sgdLogFilename)
+        else:
+            self.fwMain = FileWork(pathesDict['mainLogFilename'])
+            self.fwLogit = FileWork(pathesDict['logitLogFilename'])
+            self.fwLogitModel = FileWork(pathesDict['logitModelFilename'])
+            self.fwLassoModel = FileWork(pathesDict['fwLassoModel'])
+            self.fwSGDModel = FileWork(pathesDict['fwSGDModel'])
+            self.fwlassoLogFilename = FileWork(pathesDict['lassoLogFilename'])
+            self.fwsgdLogFilename = FileWork(pathesDict['sgdLogFilename'])
 
     def statistics(self, printStat=True, writeStat=True):
         """
@@ -64,8 +92,7 @@ class Model:
             self.fwMain.writeInFile(message=f"Percent of known age of customers in DB: {len(knownAge) / len(self.df)}")
 
     
-    def preprocessor(self, columns=['click_id', 'age', 'gender', 
-                    'var', 'isSuccsess', 'numOfCalls'], 
+    def preprocessor(self, columns=['isSuccsess', 'numOfCalls'], 
                     depVar = 'isSuccsess', withNoneValue=False, 
                     saveDataCut=0.5, test_size=0.25, returnData=False):
         """
@@ -90,13 +117,14 @@ class Model:
         if 'click_id' in columns:
             df.set_index('click_id', drop=True)
         print(len(df))
-        if not withNoneValue:
+        if not withNoneValue and 'age' in df.columns.values:
             df = df.loc[(df['age'].isna() == False) & (df['gender'] != "None")]
             if len(df) / len(self.df) <= saveDataCut:
                 raise ValueError("saveDataCut value is higer then part of dataset, "
                                     "which will be delete from original Data Frame")
         dictGender = {'m': 0, 'f':1}
-        df['gender'] = df['gender'].map(dictGender)
+        if 'gender' in df.columns.values:
+            df['gender'] = df['gender'].map(dictGender)
         columns.remove(depVar)
         x = df[columns].values
         y = df[depVar].values
@@ -134,21 +162,137 @@ class Model:
             print("__________________________________________________________________________")
             print(f"Mean accuracy on test dataset: {testScore}")
             print(f"Mean accuracy on full dataset: {fullScore}")
+            print(f"Diff in test data set: {fullScore - testScore}")
+            print(f"Test predict raiser then math expectation on {0.5 - testScore}")
+            print(f"Full predict raiser then math expectation on {0.5 - fullScore}")
+
         
         if writeStat:
             self.fwLogit.writeInFile(message="LOGIT MODEL STATS")
             self.fwLogit.writeInFile(message="__________________________________________________________________________")
             self.fwLogit.writeInFile(message=f"Mean accuracy on test dataset: {testScore}")
             self.fwLogit.writeInFile(message=f"Mean accuracy on full dataset: {fullScore}")
-        
+            self.fwLogit.writeInFile(message=f"Diff in test data set: {fullScore - testScore}")
+            self.fwLogit.writeInFile(message=f"Test predict raiser then math expectation on {testScore - 0.5}")
+            self.fwLogit.writeInFile(message=f"Full predict raiser then math expectation on {fullScore - 0.5}")
         if saveModel:
             try:
-                pickle.dump(clf, modelFilename)
+                pickle.dump(clf, self.logitModelFilename)
             except TypeError:
-                
+                check = self.fwLogitModel.createFile()
+                if not check:
+                    raise FileNotFoundError(f"Can't create/open file {self.logitModelFilename}")
             if printStat:
-                print(f"Logit model save on {modelFilename}!")
+                print(f"Logit model save on {self.logitModelFilename}!")
             if writeStat:
-                self.fwLogit.writeInFile(message=f"Logit model save on {modelFilename}!")
+                self.fwLogit.writeInFile(message=f"Logit model save on {self.logitModelFilename}!")
         
+        
+    def lassoModel(self, printStat=True, writeStat=True, saveModel=True):
+        """
+        Func for build and test Lasso model
+
+        Args:
+            printStat(bool, default: True): if True - print info
+                about model in console/terminal
+            writeStat(bool, default:True): if True - write info 
+                about model in results folder
+            logFilename(string, default:'./results/log/logit.log'):
+                path to logit model log file
+            saveModel(bool, default: True): save model on ../results/models/logit.pickle
+            modelFilename(string, default: "./results/models/logit.pickle"):
+                path to logit model pickle file (for addictive analysis)
+        """
+        try:
+            clf = Lasso(alpha=0.1).fit(self.X_train, self.y_train)
+        except ValueError:
+            raise ValueError("Find None value of dataset! Please, use"
+                                " age only (in preprocessing)")
+        testScore = clf.score(self.X_test, self.y_test)
+        fullScore = clf.score(self.full_X, self.full_y)
+
+        if printStat:
+            print("LASSO MODEL STATS")
+            print("__________________________________________________________________________")
+            print(f"Mean accuracy on test dataset: {testScore}")
+            print(f"Mean accuracy on full dataset: {fullScore}")
+            print(f"Diff in test data set: {fullScore - testScore}")
+            print(f"Test predict raiser then math expectation on {0.5 - testScore}")
+            print(f"Full predict raiser then math expectation on {0.5 - fullScore}")
+
+        
+        if writeStat:
+            self.fwLogit.writeInFile(message="LASSO MODEL STATS")
+            self.fwLogit.writeInFile(message="__________________________________________________________________________")
+            self.fwLogit.writeInFile(message=f"Mean accuracy on test dataset: {testScore}")
+            self.fwLogit.writeInFile(message=f"Mean accuracy on full dataset: {fullScore}")
+            self.fwLogit.writeInFile(message=f"Diff in test data set: {fullScore - testScore}")
+            self.fwLogit.writeInFile(message=f"Test predict raiser then math expectation on {testScore - 0.5}")
+            self.fwLogit.writeInFile(message=f"Full predict raiser then math expectation on {fullScore - 0.5}")
+        if saveModel:
+            try:
+                pickle.dump(clf, self.fwLassoModel)
+            except TypeError:
+                check = self.fwLassoModel.createFile()
+                if not check:
+                    raise FileNotFoundError(f"Can't create/open file {self.logitModelFilename}")
+            if printStat:
+                print(f"LASSO model save on {self.fwLassoModel}!")
+            if writeStat:
+                self.fwLogit.writeInFile(message=f"LASSO model save on {self.fwLassoModel}!")
+        
+        
+    def SGDModel(self, printStat=True, writeStat=True, saveModel=True):
+        """
+        Func for build and test Lasso model
+
+        Args:
+            printStat(bool, default: True): if True - print info
+                about model in console/terminal
+            writeStat(bool, default:True): if True - write info 
+                about model in results folder
+            logFilename(string, default:'./results/log/logit.log'):
+                path to logit model log file
+            saveModel(bool, default: True): save model on ../results/models/logit.pickle
+            modelFilename(string, default: "./results/models/logit.pickle"):
+                path to logit model pickle file (for addictive analysis)
+        """
+        try:
+            clf = SGDClassifier(loss="hinge", penalty="l2",
+                                    max_iter=5).fit(self.X_train, self.y_train)
+        except ValueError:
+            raise ValueError("Find None value of dataset! Please, use"
+                                " age only (in preprocessing)")
+        testScore = clf.score(self.X_test, self.y_test)
+        fullScore = clf.score(self.full_X, self.full_y)
+
+        if printStat:
+            print("SGD CLASSIFIER STATS")
+            print("__________________________________________________________________________")
+            print(f"Mean accuracy on test dataset: {testScore}")
+            print(f"Mean accuracy on full dataset: {fullScore}")
+            print(f"Diff in test data set: {fullScore - testScore}")
+            print(f"Test predict raiser then math expectation on {0.5 - testScore}")
+            print(f"Full predict raiser then math expectation on {0.5 - fullScore}")
+
+        
+        if writeStat:
+            self.fwLogit.writeInFile(message="SGD CLASSIFIER STATS")
+            self.fwLogit.writeInFile(message="__________________________________________________________________________")
+            self.fwLogit.writeInFile(message=f"Mean accuracy on test dataset: {testScore}")
+            self.fwLogit.writeInFile(message=f"Mean accuracy on full dataset: {fullScore}")
+            self.fwLogit.writeInFile(message=f"Diff in test data set: {fullScore - testScore}")
+            self.fwLogit.writeInFile(message=f"Test predict raiser then math expectation on {testScore - 0.5}")
+            self.fwLogit.writeInFile(message=f"Full predict raiser then math expectation on {fullScore - 0.5}")
+        if saveModel:
+            try:
+                pickle.dump(clf, self.fwsgdLogFilename)
+            except TypeError:
+                check = self.fwLogitModel.createFile()
+                if not check:
+                    raise FileNotFoundError(f"Can't create/open file {self.logitModelFilename}")
+            if printStat:
+                print(f"SGD CLASSIFIER save on {self.logitModelFilename}!")
+            if writeStat:
+                self.fwLogit.writeInFile(message=f"SGD CLASSIFIER save on {self.logitModelFilename}!")
